+++ bindings/clib/httpcloak.go
@@ -2524,7 +2524,7 @@ func httpcloak_free_string(str *C.char) {
 //export httpcloak_version
 func httpcloak_version() (hcRet *C.char) {
 	defer guardCharP("httpcloak_version", &hcRet)
-	return C.CString("1.6.10")
+	return C.CString("1.6.11")
 }
 
 //export httpcloak_available_presets